Add support for the EZ-Flash Parallel
It appears the Parallel already DLDI patches the boot file, so support
for this cart may *seem* redundant at first.

However, since the Parallel does this in firmware, we can't easily
update the DLDI for newer fixes. We do in fact need a different DLDI as
it fixes homebrew issues, so use dldinoap to make sure we don't get
overwritten.